// src/helper/types.ts
export interface CoinDataProps {
  __v: number;
  _id: string;
  available: string;
  base_currency: string;
  base_currency_id: string;
  buy_price: number;
  change: number;
  createdAt: string;
  high: number;
  icon_path: string;
  low: number;
  open: number;
  quote_currency: string;
  quote_currency_id: string;
  sell_price: number;
  status: string;
  type: string;
  updatedAt: string;
  volume: number;
}
export interface CoinCardProps {
  item: CoinDataProps;
  index: number;
  currency?:any;
}